Add -t and -l args

+++ update_tags.sh
@@ -1,16 +1,31 @@
 #!/bin/bash
 
+################################################################################
+# arguments
+
+v_arg=""    # version
+t_arg=""    # type (patch, minor, major)
+l_arg=false # local (skip push)
+
+################################################################################
+# functions
+
 function update_tags {
-  version=$1
+  local version=$1
+  local skip_push=$2
 
   echo
-  echo "Selected version: $version"
+  echo "New version: $version"
 
   echo "Updating tags..."
 
   echo $version
   git tag -a -m "Release $version" $version
-  git push origin $version --quiet
+
+  if [[ $l_arg == false ]]; then
+    git push origin $version --quiet
+  fi
+
   tag=$version
 
   for i in {1..2}
@@ -18,7 +33,10 @@ function update_tags {
     tag="${tag%.*}"
     echo $tag
     git tag -f -a -m "Updating tag $tag using $version" $tag
-    git push origin $tag --force --no-verify --quiet
+
+    if [[ $l_arg == false ]]; then
+      git push origin $tag --force --no-verify --quiet
+    fi
   done
 
   echo "Tags updated successfully"
@@ -34,26 +52,52 @@ function validate_tag {
   exit 1
 }
 
+################################################################################
+# main
+
+while getopts ":v:t:l" opt; do
+  case $opt in
+    v)
+      v_arg=$OPTARG
+      ;;
+    t)
+      t_arg=$OPTARG
+      ;;
+    l)
+      l_arg=true
+      ;;
+    \?)
+      echo "Invalid option: $OPTARG" >&2
+      exit 1
+      ;;
+    :)
+      echo "Option $OPTARG requires an argument." >&2
+      exit 1
+      ;;
+  esac
+done
+
+echo
 echo "Fetching tags..."
 
-git tag | xargs git tag -d > /dev/null
-git pull --tags --quiet
+if [[ $l_arg == false ]]; then
+  git tag | xargs git tag -d > /dev/null
+  git pull --tags --quiet
+fi
 
 tags=$(git tag --sort=-v:refname)
 
-while getopts ":v:" opt; do
-  if [[ $opt == "v" ]]; then
-    if [[ ${tags[@]} =~ $OPTARG ]]; then
-      echo
-      echo "Tag $OPTARG already exists"
-      exit 1
-    fi
-
-    validate_tag $OPTARG
-    update_tags $OPTARG
-    exit 0
+if [[ -n $v_arg ]]; then
+  if [[ ${tags[@]} =~ $v_arg ]]; then
+    echo
+    echo "Tag $v_arg already exists"
+    exit 1
   fi
-done
+
+  validate_tag $v_arg
+  update_tags $v_arg
+  exit 0
+fi
 
 tag=$(echo $tags | head -n 1)
 
@@ -68,7 +112,6 @@ fi
 validate_tag $tag
 
 echo "Current version: $tag"
-echo
 
 tag_components=($(echo "${tag#v}" | tr '.' ' '))
 
@@ -84,6 +127,25 @@ new_version_patch="v${major}.${minor}.${new_patch}"
 new_version_minor="v${major}.${new_minor}.0"
 new_version_major="v${new_major}.0.0"
 
+if [[ -n $t_arg ]]; then
+  case $t_arg in
+    patch)
+      update_tags $new_version_patch
+      ;;
+    minor)
+      update_tags $new_version_minor
+      ;;
+    major)
+      update_tags $new_version_major
+      ;;
+    *)
+      echo "Invalid option $t_arg"
+      ;;
+  esac
+  exit 0
+fi
+
+echo
 PS3="Select the new version: "
 
 select opt in $new_version_patch $new_version_minor $new_version_major quit; do
@@ -101,7 +163,6 @@ select opt in $new_version_patch $new_version_minor $new_version_major quit; do
       break
       ;;
     quit)
-      break
       ;;
     *)
       echo "Invalid option $REPLY"
